requirements.txt
numpy
pandas
torch
torchvision
scikit-learn
matplotlib
lime
shap
opencv-python
Pillow
flask
flask-cors
